// src/app/api/shared/lead/lead-list-response.interface.ts
export interface LeadListResponseInterface {
  data: Array<{
    id: string;
    action: boolean;
    agency_gateway_id: string;
    agency_id: string;
    agency_website_id: string;
    assign_date2: string;
    assigned_by_id: string;
    client: string;
    comment: string;
    contact_date2: string;
    contact_firstname: string;
    contact_id: string;
    contact_lastname: string;
    contact_message: string;
    contact_need_validation: string;
    created_by_id: string;
    creation_date2: string;
    duration: string;
    durationCssClass: string;
    has_email: boolean;
    has_email_status: boolean;
    has_email_privilege: boolean;
    lead_contact_id: string;
    lead_type_id: string;
    lead_type_label: string;
    main_source_id: string;
    managed_by_id: string;
    management_communication_id: string;
    management_communication_label: string;
    management_date2: string;
    manager: string;
    manager_firstname: string;
    manager_lastname: string;
    manager_initials: string;
    original_communication_id: string;
    original_communication_label: string;
    original_message: string;
    original_message_readonly: string;
    promotion_id: string;
    property_has_summary: boolean;
    property_id: string;
    property_ranking: string;
    property_agency_id: string;
    property_photo_thumb: string;
    property_photo_zoom: string;
    reference: string;
    source_label: string;
    status_id: string;
    status_label: string;
    lead_parser_error_code: string|null;
    sub_source_id: string;
    update_date2: string;
    updated_by_id: string;
    client_contact_id: string;
    client_contact_initials: string;
  }>;
  recordsTotal: string; // total number of records matching search query
}