// src/app/api/shared/lead/lead-load-response.interface.ts
export interface LeadLoadResponseInterface {
  id: string;
  lead_type_id: string;
  agency_gateway_id: string;
  agency_id: string;
  agency_website_id: string;
  assign_date: string;
  assigned_by_id: string;
  comment: string;
  contact_datetime: string;
  contact_id: string;
  contact_message: string;
  contact_need_validation: string;
  contact_string: string;
  created_by_id: string;
  creation_date: string;
  lead_contact_id: string;
  main_source_id: string;
  managed_by_id: string;
  management_communication_id: string;
  management_datetime: string;
  old_id: string;
  original_communication_id: string;
  original_message: string;
  original_message_readonly: string;
  original_portal: string;
  original_url: string;
  promotion_id: string;
  promotion_string: string;
  property_id: string;
  property_string: string;
  readonly: boolean;
  status_id: string;
  sub_source_id: string;
  lead_parser_error_code: string|null;
  lead_parser_error_data: string|null;
  success: boolean;
  has_email: boolean;
  has_email_status: boolean;
  has_email_privilege: boolean;
  update_date: string;
  updated_by_id: string;
  auto_assign_case_number: string;
}